+++ src/main/java/com/ai/boy/programming/client/OkHttpClient.java
@@ -1,13 +1,13 @@
 package com.ai.boy.programming.client;
 
 import com.intellij.openapi.diagnostic.Logger;
+import okhttp3.MediaType;
 import okhttp3.Request;
 import okhttp3.RequestBody;
 import okhttp3.Response;
 
-import java.nio.charset.StandardCharsets;
-
 import static cn.hutool.json.JSONUtil.toJsonStr;
+import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 
 /**
  * @author zhaoning
@@ -23,8 +23,8 @@ public class OkHttpClient {
             Request request = new Request.Builder()
                     .url(url)
                     .addHeader("Authorization", "Bearer " + apiKey)
-                    .addHeader("Content-Type", "application/json")
-                    .post(RequestBody.create(body.getBytes(StandardCharsets.UTF_8)))
+                    .addHeader("Content-Type", APPLICATION_JSON_VALUE)
+                    .post(RequestBody.create(MediaType.parse(APPLICATION_JSON_VALUE), body))
                     .build();
 
             okhttp3.OkHttpClient client = new okhttp3.OkHttpClient();
@@ -32,7 +32,6 @@ public class OkHttpClient {
             log.info("body: " + toJsonStr(body));
             Response response = client.newCall(request).execute();
             log.info("response: " + toJsonStr(response));
-            //System.out.println("body: " + response.body().string());
 
             return response.body().string();
         } catch (Exception e) {