+++ src/main/java/com/ai/boy/programming/setting/AppSettingsConfigurable.java
@@ -1,6 +1,7 @@
 // Copyright 2000-2023 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license.
 package com.ai.boy.programming.setting;
 
+import com.ai.boy.programming.client.impl.BedrockLlamaClient;
 import com.intellij.openapi.options.Configurable;
 import org.jetbrains.annotations.Nls;
 import org.jetbrains.annotations.Nullable;
@@ -9,6 +10,7 @@ import javax.swing.JComponent;
 import java.awt.event.ItemEvent;
 import java.awt.event.ItemListener;
 
+import static com.ai.boy.programming.enums.ProviderEnum.AMAZON;
 import static com.ai.boy.programming.factory.ModelFactory.getModelList;
 import static com.ai.boy.programming.factory.ModelFactory.getProviderList;
 
@@ -28,11 +30,6 @@ final class AppSettingsConfigurable implements Configurable {
         return "Programming Boy";
     }
 
-    @Override
-    public JComponent getPreferredFocusedComponent() {
-        return settingsComponent.getProviderCB();
-    }
-
     @Nullable
     @Override
     public JComponent createComponent() {
@@ -41,12 +38,19 @@ final class AppSettingsConfigurable implements Configurable {
     }
 
     @Override
+    public JComponent getPreferredFocusedComponent() {
+        return settingsComponent.getProviderCB();
+    }
+
+    @Override
     public boolean isModified() {
         AppSettingsState settings = AppSettingsState.getInstance();
         String selectedProvider = getSelectedProvider();
         return !selectedProvider.equals(settings.getProvider())
                 || !getSelectedModel().equals(settings.getModel(selectedProvider))
-                || !settingsComponent.getApiKey().equals(settings.getApiKey(selectedProvider));
+                || !settingsComponent.getApiKey().equals(settings.getApiKey(selectedProvider))
+                || !settingsComponent.getAccessKeyId().equals(settings.getAccessKeyId(selectedProvider))
+                || !settingsComponent.getAccessKey().equals(settings.getAccessKey(selectedProvider));
     }
 
     @Override
@@ -58,13 +62,18 @@ final class AppSettingsConfigurable implements Configurable {
         settingsState.setProvider(selectedProvider);
         settingsState.setModel(selectedProvider, selectedmodel);
         settingsState.setApiKey(selectedProvider, settingsComponent.getApiKey());
+        settingsState.setAccessKeyId(selectedProvider, settingsComponent.getAccessKeyId());
+        settingsState.setAccessKey(selectedProvider, settingsComponent.getAccessKey());
+
+        BedrockLlamaClient.resetClient();
     }
 
     @Override
     public void reset() {
         AppSettingsState settings = AppSettingsState.getInstance();
-        settingsComponent.initProvider(getProviderList(), settings.getProvider());
+        settingsComponent.initAndSelectProvider(getProviderList(), settings.getProvider());
         initModelAndApiKey();
+        initLabelVisible();
         addProviderCBListener();
     }
 
@@ -73,16 +82,31 @@ final class AppSettingsConfigurable implements Configurable {
         settingsComponent = null;
     }
 
+    private void initModelAndApiKey() {
+        // 包含了默认选择的逻辑
+        AppSettingsState settings = AppSettingsState.getInstance();
+        String selectedProvider = getSelectedProvider();
+        settingsComponent.initAndSelectModel(getModelList(selectedProvider), settings.getModel(selectedProvider));
+
+        settingsComponent.setApiKey(settings.getApiKey(selectedProvider));
+        settingsComponent.setAccessKeyId(settings.getAccessKeyId(selectedProvider));
+        settingsComponent.setAccessKey(settings.getAccessKey(selectedProvider));
+    }
+
+    private void initLabelVisible() {
+        if (getSelectedProvider().equals(AMAZON.getCode())) {
+            settingsComponent.setAccessKeyVisible();
+        } else {
+            settingsComponent.setApiKeyLabelVisible();
+        }
+    }
+
     private String getSelectedProvider() {
         return getProviderList().get(settingsComponent.getProviderCB().getSelectedIndex());
     }
 
     private String getSelectedModel() {
         String selectedProvider = getSelectedProvider();
-        return getSelectedModel(selectedProvider);
-    }
-
-    private String getSelectedModel(String selectedProvider) {
         return getModelList(selectedProvider).get(settingsComponent.getModelCB().getSelectedIndex());
     }
 
@@ -91,15 +115,8 @@ final class AppSettingsConfigurable implements Configurable {
             @Override
             public void itemStateChanged(ItemEvent e) {
                 initModelAndApiKey();
+                initLabelVisible();
             }
         });
-    }
-
-    private void initModelAndApiKey() {
-        // 包含了默认选择的逻辑
-        AppSettingsState settings = AppSettingsState.getInstance();
-        String selectedProvider = getSelectedProvider();
-        settingsComponent.initAndSelectModel(getModelList(selectedProvider), settings.getModel(selectedProvider));
-        settingsComponent.setApiKey(settings.getApiKey(selectedProvider));
     }
 }